requirements.txt
streamlit
requests
beautifulsoup4
lxml
spacy
PyPDF2
PyMuPDF
python-docx
openai
python-dotenv
